% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandf.R
\name{cleandf}
\alias{cleandf}
\title{Grouping tokenised dialogue by turn}
\arguments{
\item{data}{data frame}
}
\value{
data frame
}
\description{
This function will take in a data frame and return a new data frame that has been:
\enumerate{
\item Tokenised without stop words
\item Group the results by turn
\item Return a data frame with count, n, of each word in each turn and also the proportion of the word
}
}
\examples{
 \dontrun{
 cleandf(pedf)
}

}
