% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterwordsdf.R
\name{filterwordsdf}
\alias{filterwordsdf}
\title{Filtering dataframe to display words that are entered in a vector form}
\arguments{
\item{data}{data frame}

\item{vec}{vector of strings}
}
\value{
data frame
}
\description{
This function will filter the data frame's words according to those that are in vec
}
\examples{
 \dontrun{
 filterwordsdf(dataframe, vector)
}
}
