% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfunc.R
\name{finalfunc}
\alias{finalfunc}
\title{This function put together cleandf, filterwordsdf and graphdf.}
\arguments{
\item{data}{data frame}
}
\value{
ggplot point graph
}
\description{
Assist to clean, filter stopwords and graph data and indicated vector.
}
\examples{
\dontrun{
 finalfunc(df,vec)
 for our case we will use vector:
 "china", "dollar", "energy", "jobs", "military", "people"
 to showcase our sample data df.
 finalfunc(df, c("china", "dollar", "energy", "jobs", "military", "people"))
}
}
