% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdf.R
\name{graphdf}
\alias{graphdf}
\title{Graphing frequency of words over time of the presidential debate}
\arguments{
\item{data}{data frame}
}
\value{
ggplot point graph
}
\description{
This functions create a graph with points to illustrate the frequency that each word is mentioned per turn with points
and also adds a smooth curve to show an area of error.
}
\examples{
\dontrun{
 graphdf(dataframe)
}
}
