% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtfidf.R
\name{graphtfidf}
\alias{graphtfidf}
\title{Function to produce the tokenize dialogue by person}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
This function will graph out the tokenize dialogue by person
Option to include additional stopword with a customized vector
}
\examples{
\dontrun{
 graphtfidf(pedf, vector = add_stopw)
}
}
