% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtopnwords.R
\name{graphtopnwords}
\alias{graphtopnwords}
\title{graphtopnwords, a function which combines mostwords & finalfunc
which return result in graph}
\arguments{
\item{data}{graphtopnwords(data, n)}
}
\value{
data frame
}
\description{
Creating graph of the top "n" words
}
\examples{
\dontrun{
 graphtopnwords(df, 4)
}
}
