% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddfgithub.R
\name{loaddfgithub}
\alias{loaddfgithub}
\title{Sample Data Loading from github}
\arguments{
\item{data}{()}
}
\value{
Data Frame, df will be loaded, in this case file is loaded from https://github.com/zeonlee/math513/raw/main/df.RData
}
\description{
Sample Data from provide for assessment will be loaded from github
In this case MATH513 BD&SNV_Report_presidential_debates_2012.csv will be loaded as df
}
\examples{
\dontrun{
loaddfgithub()
}

}
