% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostwords.R
\name{mostwords}
\alias{mostwords}
\title{Finding the most words in data}
\arguments{
\item{data}{mostwords(data, n)}
}
\value{
data frame
}
\description{
Find top m words and
Sort these words according to number of times appeared
and only return top m
}
\examples{
\dontrun{
 mostwords(data, n)
}
}
