% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasethelp.R
\docType{data}
\name{pedf}
\alias{pedf}
\title{The 2012 US presidential debate dataset}
\format{
A dataframe with 3 columns, Person, Dialogue, Turn:
#' \describe{
\item{Person}{Speaker of the dialogue}
\item{Dialogue}{full dialogue from the speaker}
\item{turn}{Integer field, or sequence of the dialogues exchanged}
...
}
}
\source{
dataset from school coursework
}
\usage{
pedf
}
\description{
Dialogue data from the 2012 preseidential debate. Dataset is stored as
Rdata file and stored on https://github.com/zeonlee/math513
}
\examples{
pedf
}
\references{
dataset from school coursework
}
\keyword{US}
\keyword{debate}
\keyword{presidential}
