% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persontop.R
\name{persontop}
\alias{persontop}
\title{Group the result by person and find their top n most used words}
\arguments{
\item{data}{(n)}
}
\value{
persontop(n = n)
}
\description{
Group the result by person and find their top n most used words
}
\examples{
\dontrun{
 persontop(n)
}
}
