% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortnorder.R
\name{sortnorder}
\alias{sortnorder}
\title{Zipf's Law, Sort and arrange data}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
Sort and arrange the tokenized words by person and their tf value
}
\examples{
\dontrun{
 sortnorder(df)
}
}
