% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenwperson.R
\name{tokenwperson}
\alias{tokenwperson}
\title{tokenize the dialogue, by person and words}
\arguments{
\item{data}{()}
}
\value{
data frame
}
\description{
This function will token the dialogue and remove stopwords by person
}
\examples{
\dontrun{
 tokenwperson()
}
}
