% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenwstop.R
\name{tokenwstop}
\alias{tokenwstop}
\title{Zipf's Law, tokenize with stopwords}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
Tokenize our dialogue and removing the stop words
and calculate the tf value
}
\examples{
\dontrun{
 tokenwstop(df)
}
}
