\name{MATH513 (MDSD5 2001A, MDSE2 2001A)}
\alias{UOPMDISFEB2021-package}
\alias{UOPMDISFEB2021}
\docType{package}
\title{
\packageTitle{UOPMDISFEB2021}
}
\description{
Dataset of the three 2012 presidential debate between Obama and Romney, hosted by
Schieffer, Crowley, and Lehrer with some questions from the crowd.
%\packageDescription{UOPMDISFEB2021}
}
\format{
Sample Data Sets as well as data frame prerequisite for package:\cr
Data.frame with 2912 rows and 3 columns:\cr
Person: speaker name\cr
Dialogue: dialogue from the speaker\cr
Turn: Integer field for dialogue sequence\cr
}
\note{
Instructions on package installation available on https://github.com/zeonlee/math513\cr
Package files avalable on https://github.com/zeonlee/math513files\cr
}
\author{
\packageAuthor{UOPMDISFEB2021}

Maintainer: \packageMaintainer{UOPMDISFEB2021}
}
\references{
"10 Things You Didn't Know About Steven Chu." U.S. News & World Report, U.S. News & World Report, www.usnews.com/news/obama/articles/2008/12/30/10-things-you-didnt-know-about-steven-chu.

"October 16, 2012 Debate Transcript." CPD: October 16, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-16-2012-the-second-obama-romney-presidential-debate/.

"October 22, 2012 Debate Transcript." CPD: October 22, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-22-2012-the-third-obama-romney-presidential-debate/.

"October 3, 2012 Debate Transcript." CPD: October 3, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-3-2012-debate-transcript/
}

\keyword{ package }
