% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipffinal.R
\name{zipffinal}
\alias{zipffinal}
\title{Zipf's Law, Final Graph}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
To produce Zipf's Law graph with tokenwstop(),
sortnorder() and zipfgraph()
}
\examples{
\dontrun{
 zipffinal(df)
}
}
