% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipffinalperson.R
\name{zipffinalperson}
\alias{zipffinalperson}
\title{Zipf's Law, Final Graph by Speaker}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
To produce Zipf's Law graph with tokenwstop(),
sortnorder() and zipfgraphperson()
}
\examples{
\dontrun{
 zipffinalperson(df)
}
}
