% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfgraph.R
\name{zipfgraph}
\alias{zipfgraph}
\title{Zipf's Law, to produce a Zipf Graph}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
To produce Zipf's Law graph
}
\examples{
\dontrun{
 zipfgraph(df)
}
}
