% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfgraphperson.R
\name{zipfgraphperson}
\alias{zipfgraphperson}
\title{Zipf's Law, to produce a Zipf Graph by Speaker}
\arguments{
\item{data}{(df)}
}
\value{
data frame
}
\description{
To produce Zipf's Law graph
}
\examples{
\dontrun{
 zipfgraphperson(df)
}
}
